Extract snap-pick option tests to snap-pick-options.spec.ts
Per refactor-on-write rule: snap-picks.spec.ts reached 383 lines (over
the 360 split threshold). Option-focused describes (addSnapPickOption,
removeSnapPickOption, getSnapPickOptions, getSnapPickOptionById) now live
in a dedicated snap-pick-options.spec.ts, mirroring the existing
snap-pick-activations.spec.ts split. snap-picks.spec.ts drops to ~270
lines; the new file is ~152 lines.
